Code refactored for MotionSaver
