+++ screensaver_app/PasswordConfig.py
@@ -8,7 +8,7 @@ from PIL import Image, ImageTk, ImageDraw, ImageFont
 # Add central logging
 import sys
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from central_logger import get_logger, log_startup, log_shutdown, log_exception
+from screensaver_app.central_logger import get_logger, log_startup, log_shutdown, log_exception
 logger = get_logger('PasswordConfig')
 
 from utils.config_utils import find_user_config_path
