+++ screensaver_app/video_player.py
@@ -15,7 +15,7 @@ import getpass  # Added import
 # Add central logging
 import sys
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from central_logger import get_logger, log_startup, log_shutdown, log_exception
+from screensaver_app.central_logger import get_logger, log_startup, log_shutdown, log_exception
 logger = get_logger('VideoPlayer')
 
 from .PasswordConfig import load_config
