+++ utils/unhooks_utils/unhooks.py
@@ -12,7 +12,7 @@ import time
 import threading
 
 # Initialize central logging
-from central_logger import get_logger, log_startup, log_shutdown, log_exception
+from screensaver_app.central_logger import get_logger, log_startup, log_shutdown, log_exception
 logger = get_logger('Unhooks')
 
 # Windows-specific imports
